router/requirements.txt
aiofiles
fastapi
jinja2
python-multipart
python-networkmanager
uvicorn
